+++ src/FortitudeIO/Protocols/ORX/Serdes/Serialization/IOrxSerializer.cs
@@ -2,6 +2,6 @@
 
 public interface IOrxSerializer
 {
-    int Serialize(object message, byte[] buffer, int msgOffset, int headerOffset);
+    int Serialize(object message, byte[] buffer, nint msgOffset, int headerOffset);
     unsafe int Serialize(object message, byte* ptr, byte* endPtr);
 }
